requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
argcomplete==3.6.3
asgiref==3.10.0
asn1crypto==1.5.1
asyncpg==0.30.0
attrs==25.3.0
azure-ai-agents==1.1.0
azure-ai-inference==1.0.0b9
azure-ai-ml==1.29.0
azure-ai-projects==1.0.0
azure-ai-voicelive==1.0.0
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.35.1
azure-core-tracing-opentelemetry==1.0.0b12
azure-data-tables==12.7.0
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-mgmt-cognitiveservices==14.0.0
azure-mgmt-containerregistry==14.0.0
azure-mgmt-core==1.6.0
azure-mgmt-dns==9.0.0
azure-mgmt-keyvault==12.1.0
azure-mgmt-monitor==7.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==23.1.0
azure-mgmt-web==10.0.0
azure-monitor-opentelemetry==1.8.1
azure-monitor-opentelemetry-exporter==1.0.0b41
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
azure-storage-file-share==12.22.0
azure-storage-queue==12.13.0
bcrypt==4.1.2
black==26.1.0
blinker==1.9.0
boto3==1.40.45
botocore==1.40.45
build==1.4.0
CacheControl==0.14.3
cached-property==2.0.1
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clicksend-client==5.0.78
cloudevents==1.11.0
cloudflare==4.3.1
colorama==0.4.6
colorclass==2.2.2
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
datamodel-code-generator==0.53.0
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
fastapi==0.118.0
firebase-admin==6.9.0
fixedint==0.1.6
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
functions-framework==3.9.2
future==1.0.0
genson==1.3.0
geocoder==1.38.1
geographiclib==2.1
geoip2==5.1.0
geopy==2.4.1
google-api-core==2.25.2
google-api-python-client==2.179.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.111.0
google-cloud-appengine-admin==1.14.2
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-run==0.11.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.41.0
google-oauth==1.0.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
h3==4.3.1
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
id==1.6.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inflect==7.5.0
inflection==0.5.1
iniconfig==2.1.0
isodate==0.7.2
isort==7.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.21.3
matplotlib==3.10.6
maxminddb==2.8.2
mdurl==0.1.2
more-itertools==10.8.0
moviepy==2.2.1
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.1
msrest==0.7.1
multidict==6.6.4
mypy_extensions==1.1.0
nh3==0.3.2
numpy==2.3.3
oauthlib==3.3.1
openai==2.1.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-flask==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-urllib==0.58b0
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
passlib==1.7.4
pathable==0.4.4
pathspec==1.0.4
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
proglog==0.1.12
propcache==0.4.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.4
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydash==8.0.5
Pygments==2.19.2
PyJHora==4.5.5
PyJWT==2.10.1
pymongo==4.16.0
PyNaCl==1.6.0
pyOpenSSL==25.3.0
pyotp==2.9.0
pyparsing==3.2.5
pyproject_hooks==1.2.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pyswisseph==2.10.3.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
ratelim==0.1.6
readme_renderer==44.0
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rich==14.3.2
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.3
s3transfer==0.14.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-orm==1.2.10
sse-starlette==3.0.2
starlette==0.48.0
strictyaml==1.7.3
stripe==13.0.1
tenacity==9.1.2
terminaltables==3.1.10
timezonefinder==8.1.0
tqdm==4.67.1
twine==6.2.0
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.19.0.dev79
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.37.0
uvicorn-worker==0.4.0
watchdog==6.0.0
webauthn==2.7.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.21.0
zipp==3.23.0
zipstream-ng==1.9.0
